@@ -1,4 +1,4 @@
-class Admin < ActiveRecord::Base
+class Admin < ApplicationRecord
   has_secure_password
 
   validates_presence_of :password, :on => :create
