@@ -1,4 +1,4 @@
-class Comment < ActiveRecord::Base
+class Comment < ApplicationRecord
   belongs_to :commentable, :polymorphic => true
   belongs_to :user
 end
