@@ -1,4 +1,4 @@
-class Coupon < ActiveRecord::Base
+class Coupon < ApplicationRecord
   mount_uploader :avatar, ImageUploader
   attr_accessor :uploader_secure_token
 
