@@ -1,4 +1,4 @@
-class CouponProduct < ActiveRecord::Base
+class CouponProduct < ApplicationRecord
   belongs_to :coupon
   belongs_to :product
 end
