@@ -1,4 +1,4 @@
-class Image < ActiveRecord::Base
+class Image < ApplicationRecord
   mount_uploader :img, ImageUploader
   attr_accessor :uploader_secure_token
 
