@@ -1,4 +1,4 @@
-class Order < ActiveRecord::Base
+class Order < ApplicationRecord
   extend Enumerize  #pending 订单新建状态  settled订单结算状态  completed 订单完成支付  canceled订单取消
   enumerize :status,      in: [:paying, :pending, :settled, :confirmed, :express, :completed, :canceled], default: :pending
   enumerize :pay_method,  in: [:online, :offline], default: :online
