@@ -1,4 +1,4 @@
-class OrderProduct < ActiveRecord::Base
+class OrderProduct < ApplicationRecord
   extend Enumerize  #pending 新建状态  cooking烹饪状态  finished完成状态  canceled取消
   enumerize :status,     in: [:pending, :cooking, :finished, :canceled], default: :pending
 
